wta rule engine

+++ dto/src/main/java/com/kairos/response/dto/web/wta/ConsecutiveWorkWTATemplateDTO.java
@@ -22,7 +22,7 @@ public class ConsecutiveWorkWTATemplateDTO extends WTABaseRuleTemplateDTO {
     private long limitCount;//no of days
 
     private List<PartOfDay> partOfDays = new ArrayList<>();
-    private List<Long> plannedTimeIds = new ArrayList<>();
+    private List<BigInteger> plannedTimeIds = new ArrayList<>();
     private List<BigInteger> timeTypeIds = new ArrayList<>();
     private float recommendedValue;
     private MinMaxSetting minMaxSetting;
@@ -53,11 +53,11 @@ public class ConsecutiveWorkWTATemplateDTO extends WTABaseRuleTemplateDTO {
         this.minMaxSetting = minMaxSetting;
     }
 
-    public List<Long> getPlannedTimeIds() {
+    public List<BigInteger> getPlannedTimeIds() {
         return plannedTimeIds;
     }
 
-    public void setPlannedTimeIds(List<Long> plannedTimeIds) {
+    public void setPlannedTimeIds(List<BigInteger> plannedTimeIds) {
         this.plannedTimeIds = plannedTimeIds;
     }
