+++ kairosActivity/src/main/java/com/kairos/activity/persistence/model/wta/templates/template_types/BreaksInShiftWTATemplate.java
@@ -19,7 +19,7 @@ import java.util.List;
 public class BreaksInShiftWTATemplate extends WTABaseRuleTemplate{
     private List<BreakTemplateValue> breakTemplateValues;
     private List<BigInteger> timeTypeIds = new ArrayList<>();
-    private List<Long> plannedTimeIds = new ArrayList<>();;
+    private List<BigInteger> plannedTimeIds = new ArrayList<>();;
 
 
     public List<BigInteger> getTimeTypeIds() {
@@ -30,11 +30,11 @@ public class BreaksInShiftWTATemplate extends WTABaseRuleTemplate{
         this.timeTypeIds = timeTypeIds;
     }
 
-    public List<Long> getPlannedTimeIds() {
+    public List<BigInteger> getPlannedTimeIds() {
         return plannedTimeIds;
     }
 
-    public void setPlannedTimeIds(List<Long> plannedTimeIds) {
+    public void setPlannedTimeIds(List<BigInteger> plannedTimeIds) {
         this.plannedTimeIds = plannedTimeIds;
     }
